+++ src/main/java/de/maxhenkel/voicechat/ClientConfig.java
@@ -21,6 +21,7 @@ public class ClientConfig extends ConfigBase {
     public final ForgeConfigSpec.BooleanValue stereo;
     public final ForgeConfigSpec.BooleanValue hideIcons;
     public final ForgeConfigSpec.BooleanValue showGroupHUD;
+    public final ForgeConfigSpec.DoubleValue groupHudIconScale;
     public final ForgeConfigSpec.EnumValue<HUDIconLocation> hudIconLocation;
     public final ForgeConfigSpec.EnumValue<GroupPlayerIconOrientation> groupPlayerIconOrientation;
 
@@ -72,6 +73,9 @@ public class ClientConfig extends ConfigBase {
         showGroupHUD = builder
                 .comment("If the group HUD should be visible")
                 .define("show_group_hud", true);
+        groupHudIconScale = builder
+                .comment("The scale of the group HUD")
+                .defineInRange("group_hud_icon_scale", 2D, 0.01D, 10D);
         hudIconLocation = builder
                 .comment("The location of the HUD icons")
                 .defineEnum("hud_icon_location", HUDIconLocation.LEFT);