+++ changelog.md
@@ -1,2 +1,1 @@
-- Changed default MTU size to 1024, to make it consistent with the fabric version
-- Fixed typo in the config
+- Added group HUD scale config option